+++ protocols/MQTT/include/IoT/MQTT/DisconnectedEventSerializer.h
@@ -47,7 +47,7 @@ public:
 		
 		static const std::string REMOTING__NAMES[] = {"properties"s,"reasonCode"s,""s};
 		TypeSerializer<std::vector < IoT::MQTT::Property > >::serialize(REMOTING__NAMES[0], value.properties, ser);
-		TypeSerializer<int >::serialize(REMOTING__NAMES[1], value.reasonCode, ser);
+		TypeSerializer<int >::serialize(REMOTING__NAMES[1], static_cast<int>(value.reasonCode), ser);
 	}
 
 };
